refactor: extract aromatic-bond utilities into ichiarom.{c,h}
