+++ INCHI-1-SRC/INCHI_API/libinchi/vc14/libinchi.vcxproj
@@ -243,6 +243,7 @@
     <ClCompile Include="..\..\..\INCHI_BASE\src\ichimap2.c" />
     <ClCompile Include="..\..\..\INCHI_BASE\src\ichimap4.c" />
     <ClCompile Include="..\..\..\INCHI_BASE\src\ichinorm.c" />
+    <ClCompile Include="..\..\..\INCHI_BASE\src\ichiarom.c" />
     <ClCompile Include="..\..\..\INCHI_BASE\src\ichiparm.c" />
     <ClCompile Include="..\..\..\INCHI_BASE\src\ichiprt1.c" />
     <ClCompile Include="..\..\..\INCHI_BASE\src\ichiprt2.c" />
@@ -305,6 +306,7 @@
     <ClInclude Include="..\..\..\INCHI_BASE\src\ichimain.h" />
     <ClInclude Include="..\..\..\INCHI_BASE\src\ichimake.h" />
     <ClInclude Include="..\..\..\INCHI_BASE\src\ichinorm.h" />
+    <ClInclude Include="..\..\..\INCHI_BASE\src\ichiarom.h" />
     <ClInclude Include="..\..\..\INCHI_BASE\src\ichiring.h" />
     <ClInclude Include="..\..\..\INCHI_BASE\src\ichirvrs.h" />
     <ClInclude Include="..\..\..\INCHI_BASE\src\ichisize.h" />
