+++ INCHI-1-SRC/INCHI_EXE/inchi-1/gcc/makefile
@@ -127,6 +127,7 @@ $(P_BASE)/ichimap2.c	\
 $(P_BASE)/ichimap4.c	\
 $(P_BASE)/ichimak2.c	\
 $(P_BASE)/ichinorm.c	\
+$(P_BASE)/ichiarom.c	\
 $(P_BASE)/ichiparm.c	\
 $(P_BASE)/ichiprt1.c	\
 $(P_BASE)/ichiprt2.c	\
@@ -176,6 +177,7 @@ ichimap1.o	\
 ichimap2.o	\
 ichimap4.o	\
 ichinorm.o	\
+ichiarom.o	\
 ichiprt1.o	\
 ichiprt2.o	\
 ichiprt3.o	\
